Apply changes from cofig models
